Added CRA for react-router example
