--- examples/react-router/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <title>React Intl + React Router Example</title>
-</head>
-<body>
-    <div id="container"></div>
-    <script src="https://cdn.polyfill.io/v1/polyfill.min.js?features=Intl.~locale.en"></script>
-    <script src="react/dist/react.js"></script>
-    <script src="react-dom/dist/react-dom.js"></script>
-    <script src="bundle.js"></script>
-</body>
-</html>
